Build packages twice for node 4.5 and 6.5
./compile-docker.sh and ./compile-vagrant.sh now produce 4 binaries Debug and Release for node 4.5 and node 6.5.

User NVM under the hood, and preinstalls the required packages. The compile.sh script requires that nvm be installed for root, that node-pre-gyp, node-gyp, and node-gyp be installed, and that node-gyp has the header files for each version of node installed.

Vagrant also needed more ram for npm3 so please destroy and rebuild machines.