feat(docker): Allow JSON output of files
Currently the docker build uses a shared volume (`/out`) to place the built files on the system. This change allows us to define the output to be `JSON` and it echos a base64 encoded file (in json) on the last line.

This allows us to leverage hyper.sh (or any other docker service) to build the package and return the file without needing any security information or permissions.

This upgrades npm for docker and vagrant. It's much faster and uses less memory.

For vagrant you should rebuild your vm to get the new script and npm.