Upload to s3 (#50)
Currently the docker build uses a shared volume (`/out`) to place the built files on the system. This change allows us to PUT the tarballs to any url via the environment.

This allows us to leverage hyper.sh or AWS ECS Fargate (or any other docker service) to build and upload the files.

This upgrades npm for docker and vagrant. It's much faster and uses less memory.

For vagrant you should rebuild your vm to get the new script and npm.